[CLOTA] Add manual aplogs upload
BZ: 50990

There are more and more reports from the field and the upper management of big
battery drain. In all those cases, it is pretty difficult to get access to the
aplog. In order to solve this problem, it could be a good thing to add the
possibility to upload aplogs directly from the device in pushing a dedicated
button on CrashReport.

A new trigger was created. It's called aplog_trigger. When the user pushes the
"upload aplogs" button on CrashReport, the trigger is emited. This trigger is
received by crashlogd.
When crashlogd receives this trigger it creates n new APLOG events. N depending
on the number of packets in which the user wants to split the aplogs files.
This number can be set by persist.crashreport.packet propertie. The number
of aplogs to send is defined by persist.crashreport.aplogdepth propertie.This
propertie defines the number of aplogs sent by packet. If there is not enough
aplogs to send, only the aplogs present will be send. If packet or/and
aplogdepth properties are set to 0, 0 APLOG event will be created and no aplog
files will be sent. If propertie is not set, the default value is 3.
update code for finddir patch
Update for dev values : now 1 packet of 3 aplog
Update manage conflicts with chown

Change-Id: I244e8f30c1bc89dbd247410f0f976e3cd18bf324
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/61739
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>