[CRASHLOGD] Support a delay if necessary before uploading crashlogs
BZ: 53700

With the add of dumpstate and dropbox logs in the crashlogs directory, it could
take a long time before a crash be ready to upload. This is due to
the fact that dumpstate and dropbox take a long time to process. It's became
necessary to add a delay in order to postpone the upload of a crash.

To solve this problem crashlogd was been modified.
When a crash occurs, crashlogd runs logsystemstate service.
Since logsystemstate is a oneshot service, it can be lunched only once at a
time, so if 2 crashes occur in a short lap of time the second crash won't lunch
logsystemstate because it could be already running. To avoid that, each time a
crash occurs, crashlogd checks the logsystemstate service state and wait that
logsystemstate be stopped. When it's done, crashlogd lunchs logstystemstate and
notify CrashReport. To inform CrashReport that crashlogs are ready to upload
(logsystemstate has finished her job) a watcher on crashlog directories was
been implemented. Each time a dropbox file is created, crashlogd is informed by
the watcher with the help of the system event buffer and notify CrashReport
with the intent com.intel.crashreport.intent.CRASHLOGS_COPY_FINISHED.

Update : remove "return -1" block that close crahslogd
remove "while dumpstate" block that cause crahslogd freeze
in case of multiple crash or ANR

Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>

Change-Id: I3d72d9bae841337ea9db7cd63a0071e88e2bc247
Reviewed-on: http://android.intel.com:8080/64205
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>