AMTL: PTI logging support in AMTL
BZ: 23105

logs gsmtty to PTI devices by configuring the
ldiscs (kernel space routing)

AMTL has to add buttons to enable PTI logging in the
main window and in the Settings menu.
The following services have to be created in init.debug.rc.
The script sending SIGUSR1 signal has to be added on clovertrail
platforms.

Change-Id: Ibd0cdaa00f3f5b05b0bf889babb43c0e7f51ee69
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/64425
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>