[CRASHLOG] Create a BZ from the phone
BZ: 36209

This patch add a new feature in Phone Doctor. This feature allows the user to
report a bug and create a BZ associated with directly from the phone.

A new buffer event watcher was added. This watcher listen all the events which
occurs on data/logs/aplogs on the file bz_trigger.
When an event occurs (user ask to report a bug), a new bz_trigger file is
created. This file contains all the data about the BZ to create (summary,
description, component, type, severity and screenshot or not). Aplogs are
copied in the new bz directory. The amount of aplogs to copy depend on
persist.crashreport.aplogdepth and persist.crashreport.packet properties.
By default they are set to 1 and 3, so 3 aplogs will be copied.
If a screenshot is associated with the BZ, the screenchot is copied under
bz_screenshot.png in the BZ directory. A crashfile is also created for
this event.

Change-Id: I8ec4e266c66c84ae4bb253fca74b918c414e3474
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/67665
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>