[CRASHLOG] Improve crashlogd for  "report a bug feature"
BZ: 59555

The watcher to catch BZ trigger files isn't implemented in a good way.
Indeed, a new watcher was created to catch BZ triggers but it is associated
with /data/logs/aplogs repository as APLOGS trigger. It's useless to add a new
watcher for this repository because it's already done by the APLOGS trigger
watcher. It's possible to use the same watcher as the APLOGS.

This patch solves the problem in removing the BZ watcher and using the APLOGS
watcher instead. The code done on the APLOGS watcher has beeen modified to
integreate the BZ event treatment.

Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>

Change-Id: Ia3dbd75f6c51900b99727077ab9738746bb264fa
Reviewed-on: http://android.intel.com:8080/68167
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>