[CRASHLOG] Fix report BZ issue
BZ: 61469

The patch 67371 wasn't synchronized with the last version of the git repository
when it had been merged.
Indeed, the bz_trigger management uses the same event watcher that the
aplogs_trigger. the patch 67371 doesn't take in count this management and
breaks the "report a bug" PhoneDoctor feature.

This patch solves the problem in deleting the useless bz_trigger event watcher.

Change-Id: I27a00fbbef8214b191cd13bf54dbf35910524063
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/69556
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>