[PORT FROM MAIN]CRASHLOG: Multiple screenshots selection
BZ: 64301

Some users are confused with the screenshot selection in the BZ creation
on PhoneDoctor. They can't make the distinction between which screenshot is
selected and which are not. Moreover the possibility to attach several
screenshots to a BZ could be a good thing.

A new function was added in crashlogd in order to collect and copy all the
screenshots attached to a BZ. This function reads the bz_trigger file and for
each SCREENSHOT pair found a copy of the screenshot was done in the current
bz directory.

Change-Id: I4d1af0f79b895eaa5e50f003c877bee75b561986
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/76607
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-on: http://android.intel.com:8080/79283
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>