CRASHLOG: Creation of a fake watchdog event for fake watchdog trigger
BZ: 72888

In order to catch fake watchdog trigger in kernel_watchdog it's necessary to
catch this event on crashlogd. This trigger will generate IPANIC_FAKE and
SWWDT_RESET_FAKE.

In crashlog_check_panic function a new signature was checked in order to catch
the trigger. This signature is "EIP is at kwd_trigger_open".

Change-Id: I0cd6465ef3a713bf43ee4da0f7c8a2e8caa9349c
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/82960
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>