crashlogd : fix redundant crash logging when 'data' partition is full
BZ: 75618

When the device '/data' partition is full and a crash occurs, crashlogd logs
twice the same crash instead of once as expected.  The problem comes from a
code implementation bug when a call to history_file_write() function is
performed with input parameter 'log' to NULL.  In this case,
history_file_write() performs a crash logging (call to LOGE) whereas it is
already performed just before the call to history_file_write().

Crashes logging through LOGE have been removed from history_file_write() since
the aim of this function is to write in history_event file only.

Change-Id: Ieab5b069290875240fd7a42d5feac193aa5868c7
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/83000
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>