[PORT FROM JBMR0] AMTL : Modifications to handle logging over hsi on ctp
BZ: 63993

Amtl should provide the buttons to enable logging over hsi
on clovertrail, send the at commands to the modem and start
mts with the correct parameters.

XSIO=5 is not used at all for the logging over hsi.
Indeed, the 156MHz button is greyed out on each platform.
It should therefore be removed.

Orig-Change-Id: I132b45fdef131472471af91316f0823208349659
Change-Id: I74bf619bc415cf906b57196a543b3466f679f87f
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/83016
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>