crashlogd: Create structure to handle multiple source of events
BZ: 77732

Crashlogd is architect-ed around file monitoring, but as it tends to
address more events, we need to add new source of events.

This rework introduce I/O multiplexing with the function do_monitor();
Each source of event should provide 3 functions to be monitored :
 - <src>_init() : initialize the source
 - <src>_get_fd() : return a file descriptor which represent the new
 event trigger
 - <src>_handle() : to call when an event on the source is detected

Former do_crashlogd function is divided in this way and is renamed
file_monitor module.

Change-Id: Ie7b73b54868d8ce61303506d762495941ec65723
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Reviewed-on: http://android.intel.com:8080/84419
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>