AMTL: Enable logs in Coredump for Redhookbay
BZ: 79656

The blue button to enable logs in coredump has to be added on
the main window as well as the corresponding trace location in
the settings menu.
The coredump configuration already exists for blackbay and
lexington but has different at commands to send: at+xsio = 0
instead of at+xsio = 2 and "oct=3" instead of "oct=4"
in at+xsystrace parameters.

Change-Id: Id20892798b92ed1a8f4b560f7879dad5e3288ac3
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/86194
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>