AMTL: JAVACRASH at com.intel.amtl fix
BZ: 75736

The crash happens when Amtl is launched while modem is not ready.
An exception is thrown and caught.
While trying to destroy the activity, core.main is set to null.
However core is already null which triggers a NullPointerException.

In AmtlCore, main can be removed as finally not used.

Change-Id: I855fe9c02b99919dabb2ddad9334283695e6d93d
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/89401
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>