AMTL: Fix Activate checkbox in Settings menu
BZ: 69540

When the configuration in Settings menu corresponds to the
configuration enabled in the main window, Activate checkbox should
be checked and cannot be unchecked.
Otherwise the Activate checkbox should be unchecked. In this case,
checking it validates the configuration to apply.

Moreover the behavior of button.setChecked(true) is not reliable
as it doesn't call the onClickListener. Instead,
button.perfomClick() should be used.

The buttons corresponding to Disable BP logging have been added to
allow the user to disable logs from this menu.

Change-Id: I254ffa4bcabea56986a11c22c95591b527b10432
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/89676
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>