AMTL: Add validation when exiting with a different config
BZ: 69524

The user has now to confirm that he is really willing to
exit AMTL via the BACK button when he has choosen a different
configuration than the one running.
This applies to the Settings menu and to the main menu.

This doesn't apply to the HOME button as it cannot be overriden.

A popup has been removed because it was called twice and was
therefore displayed twice.

Change-Id: Iaac7111c9547375ce5495a06e1ee16240703922f
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/90233
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>