crashlogd: add logreader module
BZ: 33246

logreader is a new event source, allow sending event (Info/Error) when
a filtered log event is read.

Filtered log events are defined by a global array in logreader.h,
matching a range of event tags with a name and a 'crashtool'
event. Log events are reported following this pattern :

 Log            | Crashtool
 --------------------------
 EventType      | event
 Group (name)   | type
 TAG            | data0
 Message (data) | data1

See logreader.h for details.

Change-Id: Ice13ca86ca7321fac7289599ea31b11a68ebe6a8
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Reviewed-on: http://android.intel.com:8080/96647
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>