crashlogd : update DATA2 content for DUPLICATE infoevents
BZ: 89122

Infoevent DUPLICATE are generated by crashlogd when a dropbox log file is
renamed by the dropbox manager service when system date is set to an older
value. The DATA2 field of those generated infoevent currently does not contains
a relevant information.  This patch allows to retrieve the timestamp contained
in the name of the original dropbox log file and to convert it in a human
readableformat and to set it in the DUPLICATE infoevent DATA2 field allowing
then users to see clearly the original date of the crash.

Change-Id: I43dec5ec9416fb2a93d127e642a2fd6affa57713
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/93822
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>