[REVERT ME] AMTL: enable telephony stack feature
BZ: 99081

To enable the platform power management, the telephony stack should be
disabled as we are currently facing a HSIC issue.

For that purpose:
* AMTL should be updated to be able to turn on or off the telephony stack.
A platform reboot will be needed to take this modification into account. An
Android property will be set according to this configuration.
* MMGR should be updated to handle the Android property set by AMTL. If the
telephony stack is disabled, MMGR will inform its clients that the modem is out
of service. The modem remains off.

This is enabled only for saltbay products

Change-Id: I9de7bbfb23a5a9f479e614765110cf85a86fc998
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/100741
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>