CRASHLOG: No log in enhancement BZ
BZ: 97997

BZ which have "enhancement" severity don't need aplogs.
Indeed they deal with an enhancement and not with an issue.
There is no interest to have aplogs attached with.

When the user selects the severity "enhancement" in the dropbox list,
RadioButtons which displayed the number of aplogs to attach are removed.
In the bz_trigger file APLOG=0 is written to indicate to crashlogd that
no aplog are needed.

Change-Id: I469764564be1c2512068c5863f11174641e194b6
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/104058
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>