[PORT FROM MAIN]CRASHLOG: Keep BZ in history after FOTA
BZ: 97800

After a software update all events present in database are removed.
It could be interesting to keep all BZ created on the phone in database even
after a software update.

This patch corrects this problem.
BZ repository are not deleted if the BZ associated with is not present in
history_event.
BZ_logs is not reset to 0 after FOTA.

Change-Id: Ie4464d4ba5e363366b5b1c62ecd88fe51e772362
Orig-Change-Id: I6d361ad5ed543143a33b30aba4b4ad94e260f11e
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/105091
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/107410
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>