CLOTA: Issue on first call to logcat
BZ: 104606

Crashlogd fails to compute its first aplog on user build.
Indeed, aplog linked with CRASH, APLOG or BZ events are computed by crashlogd
with a call to logcat.
At first boot an aplog file is already here but with wrong right and crashlogd
can't access to it and can't overwrite it.

To fix the problem, when crashlogd computes an aplog if an aplog file is
already present it's removed.
Moreover, in compress_aplog_folder, right given to compressed files are not
good.

Update: Fix KW issues.

Change-Id: I3d2f13ef9f74c84ae1af8be1770627364a31cd62
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/105397
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>