AMTL: Modify additional traces in Settings menu
BZ: 104738

Enabling additional traces should not be possible if traces
are disabled.
Moreover the at command sent to disable aditionnal traces was
not recognized by the modem. A reboot is needed to disable
these traces.
Moreover a popup should be added when enabling mux traces to
notify the user that these traces are not persistent.
The TextView for Enabling telephony stack is removed as not used.

Change-Id: I578370321430e2e9202973e4c128a1282e6fec4f
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/108209
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>