AMTL: Disable the buttons if AlertDialog is displayed
BZ: 97339

When Modem is not ready, an AlertDialog is displayed. If the
user presses the background buttons it triggers an exception.

The AlertDialog should only be canceled by using its buttons.

"OK" should be used instead of android.R.string.ok because it
is translated when the language is changed.

Change-Id: I9534686c28957344dc4f4c6253ecf7e27102afa8
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/108198
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>