CRASHLOG: Rework PhoneDoctor for reference phones
BZ: 111208

In order to add PhoneDoctor in reference phones, some modifications have to be
done on PhoneDoctor.
Indeed, the only feature needed on reference phones is the "report a bug"
feature. Some parts of PhoneDoctor have to be removed and other have to be
kept.

To do that some parts of the code have to be common and other have to be
specific for each application(intel phone and reference phones).
common packages and specific packages have been created in each part of the
application were the code has to be separated.
NotificationReceiver was split in 2 parts (common and specific part).
The am commands used in crashlogd to notify PhoneDoctor was modified to call
the good Receiver.

Change-Id: I913ac98b515e1d6425b5e537f6e1313b077540fc
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/110420
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>