crashlogd : Issue on first call to logcat
RefBZ : 104606

Crashlogd fails to compute its first aplog on user build.
Indeed, aplog linked with CRASH, APLOG or BZ events are computed by crashlogd
with a call to logcat.
At first boot an aplog file is already here but with wrong right and crashlogd
can't access to it and can't overwrite it.

To fix the problem, when crashlogd computes an aplog if an aplog file is
already present it's removed.
Moreover, in compress_aplog_folder, right given to compressed files are not
good.

Change-Id: Iad06875961e74ec45f807cc77ec18e82daf699b5