CRASHLOG: Rework PhoneDoctor for reference phones
RefBZ: 111208

In order to add PhoneDoctor in reference phones, some modifications have to be
done on PhoneDoctor.
Indeed, the only feature needed on reference phones is the "report a bug"
feature. Some parts of PhoneDoctor have to be removed and other have to be
kept.

To do that some parts of the code have to be common and other have to be
specific for each application(intel phone and reference phones).
common packages and specific packages have been created in each part of the
application were the code has to be separated.
NotificationReceiver was split in 2 parts (common and specific part).
The am commands used in crashlogd to notify PhoneDoctor was modified to call
the good Receiver.

Change-Id: I8a1a779eb269cde3c3c9936659d4e4f45431e7c5
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>