[REVERT ME] AMTL:Set MMGR property when changing config
BZ: 120331

This fix is a workaround to integrate modem 1326.05

HSI-C driver does not go to D0i3 by default because
tracing is enabled by default in this modem FW

Workaround is to send “AT+XSYSTRACE=0” cmd via MMGR at
each boot when persist.service.mmgr.log is set to 0.
It stops when the property is set to 1.

To avoid to overwrite at+xsystrace command sent by
AMTL when applying a new logging configuration, AMTL
has to set this property to 1. When bp logging is
disabled property has to be set to 0.

Change-Id: Ic927833c709d2e121e81193b4e4d055a1ce8460e
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/116409
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>