crashlog: Modify reporting of modem coredump recovery process(Crashlogger side)
BZ: 124391

A modem coredump process could fail and in this case it could be interesting
to know why.
mmgr library was modified in this way to give the reason of a modem coredump
process failure.
This reason is given in the coredump state of mmgr_cli_coredump when the
process finishes.
If the coredump succeed the state is E_CD_SUCCEED else the state could be:
E_CD_TIMEOUT, E_CD_LINK_ERROR, E_CD_PROTOCOL_ERROR, E_CD_SELF_RESET or
E_CD_OTHER.
In each condition a MPANIC crash is generated with Data0 filled with the
coredump state. If the coredump fails, Data1 is filled with the coredump
failure reason.

During a modem coredump recovery process the device can be turned off. In this
case the modem coredump recovery process is interrupted and the MPANIC crash is
not generated. It could be interesting to be notified when a modem coredump
process was aborted by a board shutdown. For this reason a new MMGR component
event was introduced. E_MMGR_NOTIFY_CORE_DUMP event is generated when the modem
coredump process starts and E_MMG_NOTIFY_CORE_DUMP_COMPLETED is only generated
when the process is over. A mcd_processing trigger is created when the
modem coredump starts and it's only deleted when the process is over.
At the device startup the presence of mcd_processing trigger is checked.
If the trigger is here a MPANIC crash is generated with Data0 filled with
COREDUMP_ABORTED_BY_PLATFORM_SHUTDOWN.

Change-Id: I2aefd723a108126c4d2250ecd21de6cb41686fcc
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-on: http://android.intel.com:8080/120487
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>