CRASHLOG: Connect to WiFi or use MPTA for big data
BZ: 105932

Logs attached to an event are limited to a max size of 10MB.
This is due to the fact that an user could not use a WiFi network to upload his
data and that could take a lot of time to send a such amount of data.
The limit of 10MB is an issue for users which need to have more data to
understand a crash.

This limitation size was removed for bplogs files.
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>

Change-Id: I9cb31603e6a22107bd16f359a27ab8abe4c4a5dc
Reviewed-on: http://android.intel.com:8080/122624
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>