[PORT FROM MAIN] CRASHLOG: Connect to WiFi or use MPTA for big data
BZ: 105932

Logs attached to an event are limited to a max size of 10MB.
This is due to the fact that an user could not use a WiFi network to upload his
data and that could take a lot of time to send a such amount of data.
The limit of 10MB is an issue for users which need to have more data to
understand a crash.

This limitation size was removed for bplogs files.

Orig-Change-Id: I9cb31603e6a22107bd16f359a27ab8abe4c4a5dc

Change-Id: Ic0d1074ad66b803411da5f2575c491563e845a50
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>