AMTL: onModemWarmReset removal
BZ: 134631

By experience, we know that the warm reset operation is not
efficient. This warm reset operation is the less efficient
operation and the most complex to support. In several cases,
this operation is skipped (after a modem self-reset, succeed
core dump retrieval, etc). A useless context needs also to be
handled for this.

All users should remove their subscription to this event.

Change-Id: I0e7182b1db0c3144b48fb55e5918be216b46d8d5
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/128742
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Bracq, Erwan <erwan.bracq@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>