crashlogd : fix issue on event date/time computation
BZ: 136722

The crashes date/time displayed in ADB logcat logs is
never refreshed. If several crashes occur, they are all
displayed with the date/time of the 1st crash.
The date/time management has been reworked in crashlogd
to fix this issue and to make the time/date management
more efficient.

Change-Id: I14c9bec24b54c552517969cbf0c1b4681aa17087
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/130209
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>