crashlogd : Manage missing watched directories
BZ: 136350

When one or several directories nomally watched by crashlogd inotify watcher
are missing, crashlogd exits.
This behavior is modified by this patch to allow crashlogd to run even if one
or several of those directories are missing.
Moreover, in this case an infoerror is raised and an infoevent is sent for each
directory inotify watcher failed to watch.
Finally, the monitor_crashenv script has been updated to list the content of
the parent directory of directories inotify watcher failed to add to watch.

Change-Id: If00523f2855bb8660f14776b598dd976ca52e1e0
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/131430
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>