CRASHLOG: ADD RESETSRC0 and RESETSRC1 values to DATA0 and DATA1
BZ: 135243

RESETSRC0 and RESETSRC1 (or RESETIRQ1 and RESETIRQ2, depending on the CPU
board) are now copied to /sys/kernel/debug/intel_scu_osnib.
It would be a great thing to add this information in a REBOOT event.

On boot, crashlogger computes a REBOOT event. Before to add the event line
in the history_event corresponding to the REBOOT, it checks if RESET files
(RESETSRC0/1 for Tangier or RESETIRQ1/2 for other cpu) are present or not.
If files are present, crashlogger creates an eventfile in /logs/events and
writes the value of DATA3 and DATA4(RESETIRQ or RESETSRC) in it.
If files are not present, a new watcher is added on
/sys/kernel/debug/intel_scu_osnib and survey the creation of the reset files.
When the files are created, crashlogger writes their content in DATA3 and DATA4
in the REBOOT eventfile. Then it notifies PhoneDoctor that data are ready.

Change-Id: I0459f621198fd91b20c634f29735518c3bc109bc
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/132634
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>