CRASHLOG: Update APIMR reporting with new 'recovery cause' fields in MMGR message
BZ: 138438

MMGR sends a message to crashlog when a AP-Reset is done. A new API has been
implemented in which the reset causes are given in the MMGR message.

Crashlogd mmgr handle was modified in order to receive those new messages.
If there is no cause in the MMGR message, APIMR event is created with data3 set
to AP-Reset origin.
Else an APIMR event is created with data0 set to AP-Reset origin and data1 to
data5 set to cause number 1 to 5 if they exist.

Change-Id: I8326f5aa896aee53126916b95e2210e92134a0c9
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/133307
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>