CRASHLOG: Crashtool does not set correctly extra_string information when no extra string is to be set
BZ: 140353

When a modem core dump fails a reason is given in the data sent by mmgr.
Sometimes this reason could be empty and this case is not managed by crashlogd.

A check was added on the reason length given by mmgr in the data sent
to crashlogd. If this length is zero the extra_string is set to \0 else
the extra_string is set to modem core dump failure reason.

Change-Id: Id136bb243787bb395c8c8cf818ae2b3d2581feef
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/135178
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>