CRASHLOG: Add notification management for ANR,JAVACRASH and TOMBSTONE events
BZ: 144990

When a JAVACRASH, an ANR or a TOMBSTONE occurs a dumpstate/dropbox mechanism
is launched that generates a dumpstate and a dropbox file. Since this mechanism
could take a lot of time, a data ready flag is set to 0 on this event and
a notification to set this flag to 1 is done when the data become ready.
Since the new crashlogd this notification has been removed.
This patch adds again this mechanism.

Change-Id: I1780da7a28dcf4fbadb2a4bfbfebb79b76a6e1c5
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>