crashlogd: Improve "no space on /logs partition" event management
BZ: 153322

During detection of a panic event in ram console mode, a "no space on /logs partition"
event is notified in history file as soon as a copy in /logs partition fails.
This notification is raised before the crash event in history file in order to
avoid a wrong type of event notification.

Change-Id: Ic35c47f35c8f09496eda80d7f99cf8ebb1f04e7e
Signed-off-by: Lecat, AurelienX <aurelienx.lecat@intel.com>