Crashreport : Change the function "static void flush_aplog()" to non static.
trigger.c makes a call to flush_aplog, when FULL_REPORT is not defined. As the
function was static in fsutils.c the call was not visible to trigger.c.
This error was not seen until now as by default FULL_REPORT is defined.
Changing the function definition to void flush_aplog eliminates this error.
Added the function definition to fsutils.h.
Issue: APDEV-1321

Change-Id: Ib1dd2f27f1c74648f8f0dc90441debf16e16032e
Signed-off-by: Anisha Dattatraya Kulkarni <anisha.dattatraya.kulkarni@intel.com>
Reviewed-on: https://android.intel.com/164795
Reviewed-by: Zhang, Jianxun <jianxun.zhang@intel.com>
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>
Tested-by: Bergeron, Michael <michael.bergeron@intel.com>
Reviewed-by: Leung, Daniel <daniel.leung@intel.com>