AMTL: Save ConfigApplyFrag parameters in Bundle
BZ: 170475

The OS can decide to kill off Fragments independently of the
activities or at least restore them in the reverse order.
It has no knowledge of custom constructors and can only plan
on using the default, no-arg constructor.
This is why parameters should be passed through bundle as
when a fragment is restored the bundle will automatically
be restored.

Change-Id: I7c132b6cf5bc823886f380b15c7a81053736403b
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>