crashlogd: support BP logs saving in case of STATS events
BZ: 179638

BZ 171429 introduced the possibility to request saving of the BP
logs on a STATS events. This was not properly handled in
crashlogd as it tried to save the BP logs in an 'crashlog'
directory and not in a 'stats' directory.

This fixes the issue by adding a BPLOG_STATS_TYPE type.

Change-Id: I47c2cf87259dee45bff865b3130adc611696aeb5
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Signed-off-by: Riviere, Nathan <rivierex.nathan@intel.com>