[Telephony/AMTL] add cfg file for Cherrytrail
BZ: 179362

In order to enable the modem logs with AMTL, this
tool requires to have in its config_catalog folder
a file containing the mandatory parameters to be
transmitted to the modem. The file name for it
is composed by the AP CPU name and the modem FW
name. This patch creates this file, based on the
similar file for ANNIEDALE+XMM_7260 with the
required file name.

Change-Id: I4b0b66446ce859c3ce026b47d8040522ad80d7e4
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>