crashlog : failure on history_event SWUPDATE
BZ: 210631

Regardless of the presence of buildid.txt, the crashlog memory
buffer was always populated with the values from history_event.
The desired outcome is to not load the previous events when the
file is missing.
Issue not observed before, because although the buffer contained
some old values, it would only append the new fields to the file,
as long as there were still some free entries in the buffer.
Once the buffer was full, each entry triggered a whole file
rewrite, moment in which, old/obsolete values would get printed
in the log file.
Proposed solution removes history_event on startup when
buildid.txt is missing and continues to append new entries to file
until a defined value is reached, decreasing the number of full
file rewrites.

Change-Id: Ie3e934ed6febc682d7d4ddb4d98f70a3ba473e4b
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>