crashlogd: add SHA1 of mmcblk0 CID to deviceId/uuid as fallback
If SSN or emmc0_uuid driver are not available, we need a fallback
strategy to get a unique deviceId/uuid.

mmcblk0 CID is exported in a standard way through sysfs and could be
used as a replacement of emmc0_uuid diver. The SHA1 of this CID is used
to have standardised id in length and form.

DeviceId/uuid is set to 'UnknownId' as default if nothing is found,
instead of 'Medfield'.

A new function 'file_read_string' is defined is fsutils to read a one
word (non-white-space characters) from a file. Tests are added for it.

Change-Id: I7eb041205db7955e29a210c76fc0dc02564cdb84
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1993
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Reviewed-on: https://android.intel.com/275374